Cells

This module provides the Brillouin zone container.

BrillouinZoneMeshes.Cells.Cell — Type

struct Cell{T, DIM}

Container storing information of Brillouin zone. Including lattice vector, reciprocal lattice vector and their inverse; volume of unit cell and reciprocal unit cell; G vectors for extended Brillouin zone.

Lattice parameters lattice are given by a $3×3$ matrix with floating point values, where $𝐚$, $𝐛$, and $𝐜$ are given as columns.

Parameters:

T: type of data

DIM: dimension of the Brillouin zone

Members:

lattice: lattice vector

recip_lattice: reciprocal lattice vector

inv_lattice: inverse of lattice vector

inv_recip_lattice: inverse of reciprocal lattice vector

cell_volume: volume of lattice unit cell

recip_cell_volume: volume of reciprocal lattice unit cell

atoms: list of integers representing atom types

positions: # positions[i] is the location of atoms[i] in fract. coords

atom_groups`: atoms[i] == atoms[j] for all i, j in atom_group[α]

G_vector: a list of G vectors in extended Brillouin zone

symmetry: symmetry of the Brillouin zone

BrillouinZoneMeshes.Cells._compute_inverse_lattice — Method

Compute the inverse of the lattice. Require lattice to be square matrix

BrillouinZoneMeshes.Cells._compute_recip_lattice — Method

Compute the reciprocal lattice. We use the convention that the reciprocal lattice is the set of G vectors such that G ⋅ R ∈ 2π ℤ for all R in the lattice.

BrillouinZoneMeshes.Cells.default_symmetries — Method

Default logic to determine the symmetry operations to be used in the model.

BrillouinZoneMeshes.Cells.get_latvec — Method

Return I-th lattice vector of lattice. Lattice vectors are specified column-wise in lattice::Matrix.

BrillouinZoneMeshes.Cells.standard_cell — Method

function standard_cell(;
    dtype=Float64,
    lattice::AbstractMatrix,
    atoms::AbstractVector=[1,],
    positions::AbstractVector{<:AbstractVector}=[zeros(dtype, size(lattice, 1)),],
    primitive=true,
    correct_symmetry=true,
    # magnetic_moments=[],
    tol_symmetry=PointSymmetry.SYMMETRY_TOLERANCE,
    G_vector=nothing)

Returns a Cell object with crystallographic conventional cell according to the International Table of Crystallography Vol A (ITA) in case primitive=false. If primitive=true the primitive lattice is returned in the convention of the reference work of Cracknell, Davies, Miller, and Love (CDML). Of note this has minor differences to the primitive setting choice made in the ITA.