MeshMaps

This module defines MeshMap and ReducedBZMesh which provides symmetry-reduced meshes.

BrillouinZoneMeshes.MeshMaps.MeshMap — Type

struct MeshMap

Mapping from full mesh to irreducible mesh. The reduction is mainly available via symmetry operations.

Parameters:

map: mapping from full mesh to irreducible mesh. When i is the index of a point in the full mesh, map[i] is the corresponding index in reduced mesh

reduced_length: the length of reduced mesh

inv_map: inverse of map. When i is the index of a point in the reduced mesh, inv_map[i] is a list of all points corresponding to this point in the full mesh

BrillouinZoneMeshes.MeshMaps.ReducedBZMesh — Type

struct ReducedMesh{MT}

Map-reduced mesh constructed from mesh::MT with symmetry reduction.

Parameters:

mesh: bare mesh from which the reduced mesh constructed

meshmap: map from mesh to the reduced mesh